fix methods
